Basic front-end build completed
Sign-up, sign-in, change-password, and log-out are all functional and
connected to the back-end. As of now there are no bugs. I have
established the basic HTML, the api.js file, app.js file, the UI, and
all related curl-scripts. There are some non-functional buttons to be
built on as we go for adding workouts, and showing workouts. Eventually
we will add a comment ability as well for the user to input there
scores/stats.